Make invalid 'type' import assertion cause a module to fail to load
Add Modulator::ModuleTypeFromRequest to get the asserted type from
a ModuleRequest's list of import assertions, and use it to make module
creation fail if an invalid type is asserted.

Pull the ModuleType enum out of the ModuleScriptCreationParams class
so that the full class definition of ModuleScriptCreationParams  doesn't
need to be included in modulator.h.  Also make the name of the enum
values a bit less redundant, and add kInvalid to be used by
ModuleTypeFromRequest.

In preparation for subsequent import assertions changes, plumb this
module type through GetFetchedModuleScript, where it will be needed
for module map lookups since module type will be used in the module
map cache key.

Add tests for the invalid module type failure, along with tests
verifying that JS modules can still load with an empty import assertions
clause and an unsupported import assertion.

A subsequent change will require the correct type assertion to be
present to load a JSON or CSS module, and the existing JSON/CSSS module
tests will be updated to reflect the new behavior at that time.

Bug: 1132413
Change-Id: I33e3e6ae68b14d6f6c561a58c0e260a7b129d930